core/standards/css3/image-fit/js/getComputedStyle/014.html

<!doctype html>
<title>getComputedStyle, img PNG</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<img src=../../support/opera.png>
<script>
test(function() {
  var e = document.body.firstChild;
  var vals = ['none', 'fill', 'cover', 'contain', 'auto'];
  var expected = ['none', 'fill', 'cover', 'contain', 'auto'];
  for (var i = 0; i < vals.length; ++i) {
    e.style.OObjectFit = vals[i];
    assert_equals(getComputedStyle(e, '').OObjectFit, expected[i], vals[i]);
  }
});
</script>
